refactor: fix mustache lint alerts
