+++ styles.css
@@ -129,7 +129,7 @@
 }
 .local-socialcert .lsc-hero__title {
     margin: 0;
-    font-size: 1.1rem;
+    font-size: 1.3rem;
     line-height: 1.25;
     color: var(--bd-c-text);
 }
